Submission

+++ Exercise3/Animals/Special_Animals/Wolfman.cs
@@ -11,7 +11,7 @@ namespace Exercise3.Animals.Special_Animals
 {
     public class Wolfman : Wolf, IPerson //wolfman is a wolf and a person, using an interface
     {
-        public Wolfman(string name, int age, int weight, string type, bool isMale, bool isFeral, bool isPredator, bool ishostile) :
+        public Wolfman(string name, int age, double weight, string type, bool isMale, bool isFeral, bool isPredator, bool ishostile) :
         base(name, age, weight, type, isMale, isFeral, isPredator, ishostile)
         {
             IsHostile = false;  //wolfman is cursed, he is not hostile
@@ -20,5 +20,6 @@ namespace Exercise3.Animals.Special_Animals
         {
             Console.WriteLine("I am a wolfman, I can talk with you, don't kill me!");
         }
+
     }
 }
